refactor: move db to internal
